Added mangoengine as submodule.
I created a new package galah.third_party that houses it. I did this
because making changes to mangoengine and testing them with Galah was
a serious chore that usually involved the scp utility. Hopefully this
will improve things.

There's a little bit of import hackery to make any packages in
third_party's subdirectories importable.